migration cleaner massive action + improve display
git-svn-id: https://forge.indepnet.net/svn/glpi/trunk@17388 35612d4c-9d10-0410-837c-bf9dd1be60d7

+++ front/networkportmigration.form.php
@@ -80,8 +80,9 @@ if (isset($_GET["delete"])) {
          $input                     = $np->fields;
          $input['networkports_id']  = $input['id'];
          unset($input['id']);
-         $instantiation->add($input);
-         $np->delete($_POST);
+         if ($instantiation->add($input)) {
+            $np->delete($_POST);
+         }
       } else {
          Session::addMessageAfterRedirect(__('Cannot change a migration network port to an unknown one'));
       }
